Solve each and every warning

+++ ExampleProject/Source/UI/MainHUD.h
@@ -104,7 +104,7 @@ template<typename T>
 inline void MainHUD::AddPanel()
 {
 	panels_.emplace_back(std::make_unique<T>(this));
-	panelIndexMap_[typeid(T).name()] = panels_.size() - 1;
+	panelIndexMap_[typeid(T).name()] = (int)panels_.size() - 1;
 }
 
 template<typename T>
